Move classes under EjbcaRestClient
