--- signer-cloud-server/src/main/java/com/wultra/signercloud/server/ejbca/CertificateResponse.java
@@ -1,39 +1,0 @@
-/*
- * Signer Cloud
- * Copyright (C) 2025 Wultra s.r.o.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.wultra.signercloud.server.ejbca;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.Builder;
-import lombok.extern.jackson.Jacksonized;
-
-/**
- * Certificate response class.
- *
- * @author Lubos Racansky, lubos.racansky@wultra.com
- */
-@Jacksonized
-@Builder
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-record CertificateResponse(
-        String responseFormat,
-        String certificate,
-        String serialNumber,
-        String certificate_chain
-) {
-}
